Rename the constraint engine example file exports
